+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.25.0
 * Added support for using wildcard characters when referring to script
 * Updated circular script detection
+* Updated def dependencies
 
 ## 1.24.1
 * Bug fix for command split with quotes and spaced